fix: improve handlers
