+++ src/handlers/database.ts
@@ -1,0 +1,36 @@
+import { existsSync, writeFileSync } from 'fs';
+import { BanList, WebhookMessages, GivenIdList, DeleteList, VoteList } from '../typings/data/general';
+
+export function execute(
+): void {
+
+	if (existsSync('./database/bannedList.json') == false) {
+
+		writeFileSync('./database/bannedList.json', JSON.stringify(({ users: [], servers: [] }) as BanList, null, '\t'));
+	}
+
+	if (existsSync('./database/errorStacks.json') == false) {
+
+		writeFileSync('./database/errorStacks.json', JSON.stringify(({}) as WebhookMessages, null, '\t'));
+	}
+
+	if (existsSync('./database/givenIds.json') == false) {
+
+		writeFileSync('./database/givenIds.json', JSON.stringify(([]) as GivenIdList, null, '\t'));
+	}
+
+	if (existsSync('./database/toDeleteList.json') == false) {
+
+		writeFileSync('./database/toDeleteList.json', JSON.stringify(({}) as DeleteList, null, '\t'));
+	}
+
+	if (existsSync('./database/voteCache.json') == false) {
+
+		writeFileSync('./database/voteCache.json', JSON.stringify(({}) as VoteList, null, '\t'));
+	}
+
+	if (existsSync('./database/webhookCache.json') == false) {
+
+		writeFileSync('./database/webhookCache.json', JSON.stringify(({}) as WebhookMessages, null, '\t'));
+	}
+}
