+++ src/handlers/servers.ts
@@ -7,28 +7,21 @@ export async function execute(
 ): Promise<void> {
 
 	/* This updates each server to set currentlyVisiting to null. */
-	const servers = await serverModel.find();
-	console.log(`Updating ${servers.length} servers...`);
-	for (const server of servers) {
-
-		serverModel.findOneAndUpdate(
-			s => s._id === server._id,
-			(s) => { s.currentlyVisiting = null; },
-		);
-	}
+	const servers = serverModel.find();
 
 	/* For each server, it is updating the servers name in the database and adding an entry
 		to the database if the server doesn't have one. */
 	const allServers = await client.guilds.fetch();
-	for (const [, OAuth2Guild] of allServers) {
+	for (const OAuth2Guild of allServers.values()) {
 
 		try {
 
+			const server = servers.find(s => s.serverId === OAuth2Guild.id);
+			if (!server) { throw new Error('server is undefined'); }
+			if (server.name === OAuth2Guild.name) { continue; }
 			serverModel.findOneAndUpdate(
 				s => s.serverId === OAuth2Guild.id,
-				(s) => {
-					s.name = OAuth2Guild.name;
-				},
+				s => { s.name = OAuth2Guild.name; },
 			);
 		}
 		catch {
